+++ HISTORY.md
@@ -17,6 +17,7 @@
 * Change time resolution in FileOperationInfo.
 * Deleting Blob files also go through SStFileManager.
 * Remove PlainTable's store_index_in_file feature. When opening an existing DB with index in SST files, the index and bloom filter will still be rebuild while SST files are opened, in the same way as there is no index in the file.
+* Remove CuckooHash memtable.
 
 ### Bug Fixes
 * Fix a deadlock caused by compaction and file ingestion waiting for each other in the event of write stalls.